♻️ Split Dashboard Content to a component
✨ Create paginationButtons component

✨ Add paginationPageInfo component
